+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 0.0.8
+
+### Patch Changes
+
+-   Updated dependencies [575ff89]
+    -   @sap-ux/fiori-generator-shared@0.7.10
+    -   @sap-ux/ui5-library-reference-inquirer@0.3.48
+
+## 0.0.7
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.9
+-   @sap-ux/telemetry@0.5.44
+-   @sap-ux/ui5-library-reference-inquirer@0.3.47
+
+## 0.0.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.28.7
+-   @sap-ux/telemetry@0.5.43
+-   @sap-ux/ui5-library-reference-writer@0.1.35
+-   @sap-ux/fiori-generator-shared@0.7.8
+-   @sap-ux/ui5-library-reference-inquirer@0.3.46
+
+## 0.0.5
+
+### Patch Changes
+
+-   Updated dependencies [2886db3]
+    -   @sap-ux/fiori-generator-shared@0.7.7
+    -   @sap-ux/ui5-library-reference-inquirer@0.3.45
+
 ## 0.0.4
 
 ### Patch Changes